Add files via upload
Problem Statement

The Indian Premier League, or IPL, is a T20 cricket league, which was founded in 2008 and is held every year. It sees participation from both national and international players, and eight teams representing eight Indian cities compete with each other in a double round-robin format in the league stages, which are followed by playoffs. Over the years, IPL has become one of the most-watched and most attended live sporting events all over the world.

 
Results Expected

A tableau workbook containing 3 dashboards pertaining to the 3 categories - Match statistics, Player statistics and Team statistics.
Each dashboard should contain the visualisations from the sub-category worksheets. For example, in the Match statistics dashboard, the visualisations should be from the Toss Outcome vs Match Outcome, Biggest Wins and Highest Totals worksheets. Similarly, you need to prepare the dashboards for Player statistics and Team statistics.
The dashboards and the corresponding visualisations should be well-detailed.
A PPT file containing the summary of the entire visual analysis for this assignment. This ppt should contain all the visualisations that have been used along with the necessary insights and details.